Move custom field util functions to events/util.py
